+++ Mcode/+mypackage/+internal/initializePackage.m
@@ -7,7 +7,7 @@ function initializePackage
 % settings state can handle a `clear classes` gracefully.
 
 % Do not re-initialize if already initialized
-if mypackage.internal.util.getpackageappdata('initialized')
+if mypackage.internal.utils.getpackageappdata('initialized')
   return
 end
 
@@ -52,7 +52,7 @@ end
 
 % Mark library as initialized
 
-mypackage.internal.util.getpackageappdata('initialized');
+mypackage.internal.utils.setpackageappdata('initialized', true);
 
 end