+++ dev-kit/private/copyfile2.m
@@ -1,0 +1,6 @@
+function copyfile2(src, dest, varargin)
+% A version of copyfile that raises an error on failure
+
+mypackage.internal.util.copyfile2(src, dest, varargin{:});
+
+end