+++ dev-kit/private/mv.m
@@ -1,0 +1,4 @@
+function mv(source, dest)
+% A version of movefile that raises an error on failure
+mypackage.internal.util.mv(source, dest);
+end