+++ dev-kit/private/reporoot.m
@@ -1,0 +1,4 @@
+function out = reporoot
+% The root dir of the myproject repo
+out = string(fileparts(fileparts(fileparts(mfilename('fullpath')))));
+end