+++ dev-kit/private/withcd.m
@@ -1,0 +1,6 @@
+function out = withcd(dir)
+% Temporarily change to a new directory
+
+out = mypackage.internal.util.withcd(dir);
+
+end