+++ doc-src-gh-pages-raw/make_doc.m
@@ -1,0 +1,14 @@
+function make_doc
+% Build these doc sources into the final doc/ directory
+
+%#ok<*STRNU>
+
+import mypackage.internal.util.*
+
+RAII.cd = withcd(fileparts(mfilename('fullpath')));
+rmdir2('../doc', 's');
+mkdir2('../doc');
+copyfile2('*.*', '../doc');
+
+end
+